requirements.txt
flask
boto3
flask_sqlalchemy
flask_script
flask_migrate
gunicorn
psycopg2
python-telegram-bot
requests
